Fix for missing UI Components
InputDate, PickList, TextArea, and FieldValue were not being displayed due to an additional if condition. I need someone more familiar with UI changes to validate that removing this if condition won't cause an issue. I'm not familiar with the visible property on the param config and couldn't find where it was being set or used. Is it still relevant?